# openapi.yaml
openapi: 3.1.0
info:
  title: ORP API
  version: 1.0.0
  description: |
    REST API for the ORP (Object Relationship Platform) Data Fusion engine.
    Provides entity CRUD, ORP-QL queries, graph traversal, real-time monitoring,
    connector management, and alert rules for maritime and multi-domain operations.

    ## Authentication
    All endpoints require either a Bearer JWT (OIDC) or an API key (`X-API-Key` header).

    ## Pagination
    All list endpoints support `page`, `limit` (max 1000), `sort_by`, and `sort_order` query parameters.

    ## Rate Limiting
    Default: 1000 req/sec per client API key. Exceeding the limit returns `429 Too Many Requests`.
  contact:
    name: ORP Team
    url: https://github.com/orproject/orp
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:9090/api/v1
    description: Local development
  - url: https://orp.example.com/api/v1
    description: Production
security:
  - BearerAuth: []
  - ApiKeyAuth: []
...
paths:
...
  /entities:
    get:
      operationId: listEntities
      summary: List all entities
      description: |
        Returns a paginated list of all entities with optional filtering by type, tags,
        and creation/update timestamps.
      tags: [entities]
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
        - name: type
          in: query
          description: Filter by entity type (e.g. Ship, Port, WeatherSystem)
          schema:
            type: string
          example: Ship
        - name: tags
          in: query
          description: Filter by tags (OR logic). Repeat param for multiple tags.
          schema:
            type: array
            items:
              type: string
          style: form
          explode: true
          example: [container, active]
        - name: created_after
          in: query
          description: Return entities created after this ISO 8601 timestamp
          schema:
            type: string
            format: date-time
        - name: created_before
          in: query
          schema:
            type: string
            format: date-time
        - name: updated_after
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Paginated entity list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EntityListResponse'
              examples:
                ships:
                  summary: List of ships
                  value:
                    data:
                      - id: ship-imo-1234567
                        type: Ship
                        name: Maersk Seatrade
                        tags: [container, international, active]
                        properties:
                          imo: "1234567"
                          mmsi: "219018200"
                          call_sign: OMCN
                          type: Container ship
                          length: 366.0
                          beam: 48.8
                          speed: 18.5
                          course: 225.0
                          destination: Rotterdam
                        geometry:
                          type: Point
                          coordinates: [3.2847, 51.9225]
                        confidence: 0.99
                        freshness:
                          updated_at: '2026-03-26T10:28:45Z'
                          checked_at: '2026-03-26T10:30:00Z'
                        created_at: '2026-03-20T14:00:00Z'
                        updated_at: '2026-03-26T10:28:45Z'
                    pagination:
                      page: 1
                      limit: 10
                      total_count: 2847
                      total_pages: 285
                      has_next: true
                      has_prev: false
                empty:
                  summary: Empty result set
                  value:
                    data: []
                    pagination:
                      page: 1
                      limit: 100
                      total_count: 0
                      total_pages: 0
                      has_next: false
                      has_prev: false
                ports:
                  summary: List of ports
                  value:
                    data:
                      - id: port-rdam-01
                        type: Port
                        name: Port of Rotterdam
                        tags: [major, european, container]
                        properties:
                          country: Netherlands
                          iata_code: NLRTM
                          total_teu: 14000000
                          congestion: 0.65
                        geometry:
                          type: Point
                          coordinates: [4.2706, 51.9289]
                        confidence: 1.0
                        freshness:
                          updated_at: '2026-03-26T08:00:00Z'
                          checked_at: '2026-03-26T10:00:00Z'
                        created_at: '2026-01-01T00:00:00Z'
                        updated_at: '2026-03-26T08:00:00Z'
                    pagination:
                      page: 1
                      limit: 100
                      total_count: 512
                      total_pages: 6
                      has_next: true
                      has_prev: false
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      operationId: createEntity
      summary: Create or upsert an entity
      description: |
        Ingest a new entity. If an entity with the same `id` already exists, its
        properties are merged (upsert behaviour).
      tags: [entities]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEntityRequest'
            examples:
              ship:
                summary: Create a ship entity
                value:
                  id: ship-imo-1234567
                  type: Ship
                  name: Maersk Seatrade
                  tags: [container, international]
                  properties:
                    imo: "1234567"
                    mmsi: "219018200"
                    call_sign: OMCN
                    type: Container ship
                    speed: 18.5
                    course: 225.0
                    destination: Rotterdam
                  geometry:
                    type: Point
                    coordinates: [3.2847, 51.9225]
                  confidence: 0.99
                  source: ais-feed-1
              port:
                summary: Create a port entity
                value:
                  id: port-rdam-01
                  type: Port
                  name: Port of Rotterdam
                  tags: [major, european, container]
                  properties:
                    country: Netherlands
                    iata_code: NLRTM
                    total_teu: 14000000
                    congestion: 0.65
                    container_terminals: 10
                  geometry:
                    type: Point
                    coordinates: [4.2706, 51.9289]
                  confidence: 1.0
                  source: openport-database
              weather:
                summary: Create a weather system entity
                value:
                  id: weather-storm-1
                  type: WeatherSystem
                  name: Storm Cell Alpha
                  tags: [storm, active]
                  properties:
                    severity: high
                    wind_speed_knots: 45
                    radius_km: 250
                  geometry:
                    type: Point
                    coordinates: [2.5, 51.0]
                  confidence: 0.87
                  source: weather-service
      responses:
        '201':
          description: Entity created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Entity'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
  /entities/search:
    get:
      operationId: searchEntities
      summary: Advanced entity search
      description: |
        Multi-criterion search with geospatial predicates, full-text search,
        property filters, and bounding-box / radius queries.
      tags: [entities]
      parameters:
        - name: type
          in: query
          schema:
            type: string
          example: Ship
        - name: near
          in: query
          description: "Geospatial point radius filter. Format: lat,lon,radius_km"
          schema:
            type: string
          example: "51.9225,4.2706,50"
        - name: min_lat
          in: query
          schema:
            type: number
            format: double
        - name: min_lon
          in: query
          schema:
            type: number
            format: double
        - name: max_lat
          in: query
          schema:
            type: number
            format: double
        - name: max_lon
          in: query
          schema:
            type: number
            format: double
        - name: text_search
          in: query
          description: Full-text search on name and properties
          schema:
            type: string
          example: Maersk
        - name: tags
          in: query
          schema:
            type: array
            items:
              type: string
          style: form
          explode: true
        - name: created_after
          in: query
          schema:
            type: string
            format: date-time
        - name: created_before
          in: query
          schema:
            type: string
            format: date-time
        - name: updated_after
          in: query
          schema:
            type: string
            format: date-time
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/EntityListResponse'
                  - type: object
                    properties:
                      search_time_ms:
                        type: integer
                        description: Search execution time in milliseconds
              examples:
                nearRotterdam:
                  summary: Ships within 50km of Rotterdam
                  value:
                    data:
                      - id: ship-imo-1234567
                        type: Ship
                        name: Maersk Seatrade
                        tags: [container]
                        properties:
                          speed: 18.5
                          course: 225.0
                        geometry:
                          type: Point
                          coordinates: [3.2847, 51.9225]
                        confidence: 0.99
                        freshness:
                          updated_at: '2026-03-26T10:28:45Z'
                          checked_at: '2026-03-26T10:30:00Z'
                        created_at: '2026-03-20T14:00:00Z'
                        updated_at: '2026-03-26T10:28:45Z'
                    pagination:
                      page: 1
                      limit: 100
                      total_count: 23
                      total_pages: 1
                      has_next: false
                      has_prev: false
                    search_time_ms: 12
                textSearch:
                  summary: Full-text search results
                  value:
                    data:
                      - id: ship-imo-1234567
                        type: Ship
                        name: Maersk Seatrade
                        tags: [container]
                        properties: {}
                        confidence: 0.99
                        freshness:
                          updated_at: '2026-03-26T10:28:45Z'
                          checked_at: '2026-03-26T10:30:00Z'
                        created_at: '2026-03-20T14:00:00Z'
                        updated_at: '2026-03-26T10:28:45Z'
                    pagination:
                      page: 1
                      limit: 100
                      total_count: 5
                      total_pages: 1
                      has_next: false
                      has_prev: false
                    search_time_ms: 8
                noResults:
                  summary: Search with no matches
                  value:
                    data: []
                    pagination:
                      page: 1
                      limit: 100
                      total_count: 0
                      total_pages: 0
                      has_next: false
                      has_prev: false
                    search_time_ms: 3
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /entities/{id}:
    parameters:
      - $ref: '#/components/parameters/EntityIdParam'

    get:
      operationId: getEntity
      summary: Get a single entity
      description: |
        Returns the complete entity record including all properties, geometry,
        relationships (both directions), and change history.
      tags: [entities]
      responses:
        '200':
          description: Entity with full details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EntityDetail'
              examples:
                ship:
                  summary: Ship with relationships
                  value:
                    id: ship-imo-1234567
                    type: Ship
                    name: Maersk Seatrade
                    tags: [container, international, active]
                    properties:
                      imo: "1234567"
                      mmsi: "219018200"
                      speed: 18.5
                      course: 225.0
                      destination: Rotterdam
                    geometry:
                      type: Point
                      coordinates: [3.2847, 51.9225]
                    confidence: 0.99
                    freshness:
                      updated_at: '2026-03-26T10:28:45Z'
                      checked_at: '2026-03-26T10:30:00Z'
                    created_at: '2026-03-20T14:00:00Z'
                    updated_at: '2026-03-26T10:28:45Z'
                    relationships:
                      outgoing:
                        - relationship_type: HEADING_TO
                          target_id: port-rdam-01
                          target_type: Port
                          target_name: Port of Rotterdam
                          properties:
                            eta: '2026-03-28T14:00:00Z'
                            cargo: containers
                        - relationship_type: OPERATED_BY
                          target_id: org-maersk
                          target_type: Organization
                          target_name: A.P. Moller - Maersk
                      incoming:
                        - relationship_type: THREATENS
                          source_id: weather-storm-1
                          source_type: WeatherSystem
                          source_name: Storm Cell Alpha
                          properties:
                            distance_km: 180
                            severity: high
                    history:
                      - timestamp: '2026-03-26T10:28:45Z'
                        changed_properties:
                          speed: 18.5
                          course: 225.0
                        source: ais-feed-1
                notFound:
                  summary: Entity not found response
                  value:
                    error:
                      code: NOT_FOUND
                      status: 404
                      message: Entity not found
                      details:
                        id: ship-unknown-999
                        type: Entity
                      timestamp: '2026-03-26T10:30:45Z'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    put:
      operationId: updateEntity
      summary: Update (merge) entity properties
      description: |
        Merges the supplied fields into the existing entity. Properties not included
        in the request body are left unchanged (partial update / merge semantics).
      tags: [entities]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEntityRequest'
            examples:
              updatePosition:
                summary: Update ship position and speed
                value:
                  properties:
                    speed: 19.2
                    course: 230.0
                    destination: Antwerp
                  geometry:
                    type: Point
                    coordinates: [3.5, 51.95]
              updateTags:
                summary: Replace tags
                value:
                  tags: [container, diverted]
              updateConfidence:
                summary: Update confidence score
                value:
                  confidence: 0.75
      responses:
        '200':
          description: Updated entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Entity'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'

    delete:
      operationId: deleteEntity
      summary: Delete entity
      description: |
        Soft-deletes the entity. The record is marked as deleted and excluded from
        normal queries, but history is preserved for auditing.
      tags: [entities]
      responses:
        '204':
          description: Entity deleted successfully (no response body)
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /entities/{id}/relationships:
    parameters:
      - $ref: '#/components/parameters/EntityIdParam'

    get:
      operationId: getEntityRelationships
      summary: Get entity relationships
      description: |
        Returns all incoming and outgoing relationships for an entity, along with
        aggregate statistics.
      tags: [relationships]
      responses:
        '200':
          description: Relationship graph for the entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RelationshipsResponse'
              examples:
                shipRelationships:
                  summary: Ship relationships
                  value:
                    entity_id: ship-imo-1234567
                    entity_type: Ship
                    outgoing:
                      - id: rel-001
                        type: HEADING_TO
                        target_id: port-rdam-01
                        target_type: Port
                        target_name: Port of Rotterdam
                        properties:
                          eta: '2026-03-28T14:00:00Z'
                          cargo: containers
                        strength: 0.97
                        created_at: '2026-03-26T08:00:00Z'
                        updated_at: '2026-03-26T10:00:00Z'
                    incoming:
                      - id: rel-002
                        type: THREATENS
                        source_id: weather-storm-1
                        source_type: WeatherSystem
                        source_name: Storm Cell Alpha
                        properties:
                          distance_km: 180
                          severity: high
                        strength: 0.88
                        created_at: '2026-03-26T06:00:00Z'
                        updated_at: '2026-03-26T10:00:00Z'
                    statistics:
                      total_outgoing: 2
                      total_incoming: 1
                emptyRelationships:
                  summary: Entity with no relationships
                  value:
                    entity_id: port-new-01
                    entity_type: Port
                    outgoing: []
                    incoming: []
                    statistics:
                      total_outgoing: 0
                      total_incoming: 0
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /monitors/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: Monitor ID
        schema:
          type: string

    put:
      operationId: updateMonitor
      summary: Update an alert monitor
      tags: [monitors]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMonitorRequest'
            examples:
              disable:
                summary: Disable a monitor
                value:
                  enabled: false
              updateCooldown:
                summary: Adjust cooldown period
                value:
                  cooldown_seconds: 600
      responses:
        '200':
          description: Updated monitor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Monitor'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    delete:
      operationId: deleteMonitor
      summary: Delete an alert monitor
      tags: [monitors]
      responses:
        '204':
          description: Monitor deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  # ── ALERTS ──────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: OIDC JWT token obtained from the /auth/login flow
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key with scoped permissions (prefix orpk_prod_)
  parameters:
...
    PageParam:
      name: page
      in: query
      description: Page number (1-indexed)
      schema:
        type: integer
        minimum: 1
        default: 1
      example: 1
    LimitParam:
      name: limit
      in: query
      description: Items per page (max 1000)
      schema:
        type: integer
        minimum: 1
        maximum: 1000
        default: 100
      example: 50
    SortByParam:
      name: sort_by
      in: query
      description: Field name to sort by
      schema:
        type: string
      example: created_at
    SortOrderParam:
      name: sort_order
      in: query
      description: Sort direction
      schema:
        type: string
        enum: [asc, desc]
        default: desc
...
  responses:
...
    Conflict:
      description: Conflict — resource already exists or state conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...